# Use output size estimation to speed up serialization