--- a/src/blockdata/block.rs
+++ b/src/blockdata/block.rs
@@ -532,9 +532,9 @@ mod tests {
 #[cfg(bench)]
 mod benches {
     use super::Block;
-    use crate::EmptyWrite;
     use crate::consensus::{deserialize, Encodable, Decodable};
     use test::{black_box, Bencher};
+    use crate::prelude::sink;
 
     #[bench]
     pub fn bench_stream_reader(bh: &mut Bencher) {
@@ -571,7 +571,7 @@ mod benches {
         let block: Block = deserialize(&raw_block[..]).unwrap();
 
         bh.iter(|| {
-            let size = block.consensus_encode(&mut EmptyWrite);
+            let size = block.consensus_encode(&mut sink());
             black_box(&size);
         });
     }