--- a/src/blockdata/locktime/mod.rs
+++ b/src/blockdata/locktime/mod.rs
@@ -1,0 +1,7 @@
+// Rust Bitcoin Library - Written by the rust-bitcoin developers.
+// SPDX-License-Identifier: CC0-1.0
+
+//! Provides absolute locktime.
+//!
+
+pub mod absolute;