--- a/src/blockdata/locktime/mod.rs
+++ b/src/blockdata/locktime/mod.rs
@@ -1,7 +1,8 @@
 // Rust Bitcoin Library - Written by the rust-bitcoin developers.
 // SPDX-License-Identifier: CC0-1.0
 
-//! Provides absolute locktime.
+//! Provides absolute and relative locktimes.
 //!
 
 pub mod absolute;
+pub mod relative;