--- a/src/blockdata/transaction.rs
+++ b/src/blockdata/transaction.rs
@@ -973,6 +973,10 @@ impl Encodable for Transaction {
         len += self.lock_time.consensus_encode(w)?;
         Ok(len)
     }
+
+    fn serialized_len_est(&self) -> usize {
+        64 + self.input.len() * 192 + self.output.len() * 48
+    }
 }
 
 impl Decodable for Transaction {