This is a continuation of experimenting/discussion started in #1224 . No matter the disagreements, any pre-allocating is going to speed things up so we should decide on something.

I left the existing PRs from #1224 intact, as `STATIC_SERIALIZED_LEN` is used
by the new `serialized_len_est` when available (because it's better to be precise
when it's free).

Now some benchmarks of the two functions that matter most.

For baseline, `master` with the test changed to `consensus::serialize` like everywhere else:

```
test blockdata::block::benches::bench_block_serialize                      ... bench:     205,827 ns/iter (+/- 10,088)
test blockdata::transaction::benches::bench_transaction_serialize          ... bench:         184 ns/iter (+/- 8)

test blockdata::transaction::benches::bench_transaction_serialize_prealloc ... bench:          25 ns/iter (+/- 1)
```
That 25ns/iter seems like a prefect case (where the buffer was pre-alloced to the correct size ahead of time). Notably this is not achievable in rest of the tests, because this function does not allocate even a single time in the benchmark loop, reusing the same buffer every time.

Plain #1224:

```
test blockdata::block::benches::bench_block_serialize                      ... bench:     205,305 ns/iter (+/- 11,689)
test blockdata::transaction::benches::bench_transaction_serialize          ... bench:          48 ns/iter (+/- 2)
```

This PR:

```
test blockdata::block::benches::bench_block_serialize                      ... bench:     203,443 ns/iter (+/- 10,727)
test blockdata::transaction::benches::bench_transaction_serialize          ... bench:          35 ns/iter (+/- 1)
```

From all observations here it seem the cost of any (re)allocation is around 10ns on my system, at least for smallish sizes. Calculating the exact space for the tx used:

```
test blockdata::transaction::benches::bench_transaction_serialize_logic    ... bench:           8 ns/iter (+/- 0)                                                          
```

which is in the ballpark of the time saved, by using estimation. I'm not sure why exactly it's not equal - potentially it's the effect of some extra icache/branch predictor misses.
 
While the benchmarking shows that estimating instead of being precise seems to be around 10ns faster for this particular transaction, it's important to remember that it potentially wastes some memory. How much depends on the transaction and the estimation itself. I picked some numbers by just looking at handful of samples from the block explorer. Precise allocation does not waste any memory up to sizes of 512. After which it wastes between 0% to 50%, due to the way Rust grows collection size. It will also potentially reallocate multiple times for big txes. Estimation (if not messed up) has a chance of statistically wasting less, as it should over-estimate slightly, and rarely doing 1 extra allocation when it was wrong.

Serializing block is also slightly faster, and might benefit in the same way as big transactions by statistically over-allocating less (when correct), and reallocating fewer times.

Estimation can just use precise counting when it makes sense.